CategoryAuth

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Models\Product;
 
@@ -10,6 +11,9 @@ class CategoryController extends Controller
 {
     public function index()
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         return view('dashboard.categories.index', [
            'categories' => Category::all()
         ]);
@@ -17,6 +21,9 @@ class CategoryController extends Controller
 
     public function create(Category $category)
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         return view('dashboard.categories.create', [
             'category' => $category
         ]);
@@ -24,6 +31,9 @@ class CategoryController extends Controller
 
     public function store()
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         $attributes = request()->validate([
             'name'=>'required',
 
@@ -35,11 +45,17 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         return view('dashboard.categories.edit', ['category'=>$category]);
     }
 
     public function update(Category $category)
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         $attributes = request()->validate([
             'name'=>'required',
 
@@ -51,7 +67,11 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if (! Gate::any(['superadmin', 'admin'])) {
+            abort(403);
+        }
         Product::where('category_id', $category->id)->update(['category_id' => null]);
+
         $category->delete();
         return redirect('/dashboard/categories/');
